feat: add "done" state to visualization
Allows to mark nodes in the visualization as "done" by double-clicking
on them. A "done" node will be shown with a green background.

Double-clicking on a "done" node will reset them to their initial state.

Note: this "done" state is not persisted in LocalStorage, it is only
saved in the visualization state. This means that if you e.g. reload the
page, all nodes will loose their "done" state.